+++ CHANGELOG.md
@@ -14,6 +14,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - **Magic Mirror 2.2.0 is now required to run this module (Nunjucks support required)**
 - Retry to connect to endpoint in integration tests up to two times
 - Reduced range of temperature icons from 8 to 7 degrees
+- Moved current text and windspeed out of the temperature div
 
 
 ## [1.2.0] - 2018-11-14